msm: msm_bus: Enable msm-bus test to build all qcom targets
Enable building msm_bus kernel test module for all the
qualcomm targets.

Resolve compilation issue on 64-bit target.

+++ msm_bus/msm_bus_test.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, The Linux Foundation. All rights reserved.
+ * Copyright (c) 2013, 2016 The Linux Foundation. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
@@ -87,7 +87,7 @@ struct msm_bus_test_scale_pdata {
 struct msm_bus_test_cldata {
 	struct msm_bus_test_scale_pdata pdata;
 	uint32_t clid;
-	uint32_t pdatah;
+	size_t pdatah;
 };
 
 struct msm_bus_test_update_req_data {
